Add files via upload
The most difficult thing when you use multiple accounts is to generate a strong and different password. The strong password would contain alphanumeric, numbers, alphabets and characters. The best use of Python is to build a project which could generate quick, different, random and customized passwords for your multiple accounts.